refactor filter types in preparation of accepting multiple params change at the same time
