+++ libs/openchallenges/org-search/src/lib/org-search-filter-panels.ts
@@ -1,0 +1,19 @@
+import { FilterPanel } from '@sagebionetworks/openchallenges/ui';
+import {
+  challengeContributionRolesFilter,
+  organizationCategoriesFilter,
+} from './org-search-filters';
+
+export const challengeContributionRolesFilterPanel: FilterPanel = {
+  query: 'challengeContributionRoles',
+  label: 'Contribution Role',
+  options: challengeContributionRolesFilter,
+  collapsed: false,
+};
+
+export const organizationCategoriesFilterPanel: FilterPanel = {
+  query: 'categories',
+  label: 'Category',
+  options: organizationCategoriesFilter,
+  collapsed: false,
+};
