--- libs/openchallenges/org-search/src/lib/org-search-filters-values.ts
@@ -1,34 +1,0 @@
-import { FilterValue } from '@sagebionetworks/openchallenges/ui';
-
-export const challengeContributionRolesFilterValues: FilterValue[] = [
-  {
-    value: 'challenge_organizer',
-    label: 'Challenge Organizer',
-  },
-  {
-    value: 'data_contributor',
-    label: 'Data Contributor',
-  },
-  {
-    value: 'sponsor',
-    label: 'Sponsor',
-  },
-];
-
-export const organizationCategoriesFilterValues: FilterValue[] = [
-  {
-    value: 'featured',
-    label: 'Featured',
-  },
-];
-
-export const organizationSortFilterValues: FilterValue[] = [
-  {
-    value: 'relevance',
-    label: 'Relevance',
-  },
-  {
-    value: 'challenge_count',
-    label: 'Challenge Count',
-  },
-];
