+++ libs/openchallenges/org-search/src/lib/org-search.component.html
@@ -1,4 +1,4 @@
-<main class="base mat-typography">
+<!-- <main class="base mat-typography">
   <section id="search-top">
     <div id="title" class="row">
       <div class="col col-10">
@@ -84,4 +84,4 @@
   [privacyPolicyUrl]="privacyPolicyUrl"
   [termsOfUseUrl]="termsOfUseUrl"
   [apiDocsUrl]="apiDocsUrl"
-/>
+/> -->
