+++ libs/openchallenges/ui/src/index.ts
@@ -5,8 +5,8 @@ export * from './lib/challenge-card/challenge-card.component';
 export * from './lib/challenge-card/mock-challenges';
 export * from './lib/challenge-card/mock-platforms';
 export * from './lib/checkbox-filter/checkbox-filter.component';
-export * from './lib/checkbox-filter/filter-value.model';
 export * from './lib/checkbox-filter/filter.model';
+export * from './lib/checkbox-filter/filter-panel.model';
 export * from './lib/footer/footer.component';
 export * from './lib/navbar/navbar-section';
 export * from './lib/navbar/navbar.component';
