--- libs/openchallenges/ui/src/lib/checkbox-filter/filter-value.model.ts
@@ -1,10 +1,0 @@
-import { DateRange } from './data-range';
-
-export interface FilterValue {
-  /* The value of the filter. */
-  value: DateRange | string | string[] | number | undefined;
-  /* The display name of the filter value. */
-  label: string | undefined;
-  /* The avatar url of the filter value. */
-  avatarUrl?: string | null;
-}
