+++ libs/openchallenges/ui/src/lib/checkbox-filter/filter.model.ts
@@ -1,14 +1,12 @@
-import { FilterValue } from './filter-value.model';
+import { DateRange } from './data-range';
+
+export type FilterValue = DateRange | string | number | undefined;
 
 export interface Filter {
-  /* The property name used to to query the data  */
-  query: string;
-  /* The label name of the filter  */
-  label: string;
-  /* The values of the filter. */
-  values: FilterValue[];
-  /* Whether this panel of filter is collapsed. */
-  collapsed: boolean;
-  /* Whether to show the avatar of filter value (now only applicable to dropdown filter). */
-  showAvatar?: boolean;
+  /* The value of the filter. */
+  value: FilterValue | FilterValue[];
+  /* The display name of the filter value. */
+  label: string | undefined;
+  /* The avatar url of the filter value. */
+  avatarUrl?: string | null;
 }
